feat: Separate hooks of home page
